Move CSV tools out of massedit, small structural changes
